Move Voidform to Spells
